fix(chore): put trusted into each tab
first draft
