+++ src/gui/MainWindow.cc
@@ -183,115 +183,7 @@
 #endif
 
 #ifdef ENABLE_PYTHON
-#include "nettle/base64.h"
-#include "nettle/sha2.h"
-#include "nettle/version.h"
 #include "python/python_public.h"
-
-std::string SHA256HashString(std::string aString)
-{
-  uint8_t digest[SHA256_DIGEST_SIZE];
-  sha256_ctx sha256_ctx;
-
-  sha256_init(&sha256_ctx);
-  sha256_update(&sha256_ctx, aString.length(), (uint8_t *)aString.c_str());
-#if NETTLE_VERSION_MAJOR >= 4
-  sha256_digest(&sha256_ctx, digest);
-#else
-  sha256_digest(&sha256_ctx, SHA256_DIGEST_SIZE, digest);
-#endif
-
-  base64_encode_ctx base64_ctx;
-  char digest_base64[BASE64_ENCODE_LENGTH(SHA256_DIGEST_SIZE) + 1];
-  memset(digest_base64, 0, sizeof(digest_base64));
-
-  base64_encode_init(&base64_ctx);
-  base64_encode_update(&base64_ctx, digest_base64, SHA256_DIGEST_SIZE, digest);
-  base64_encode_final(&base64_ctx, digest_base64);
-  return digest_base64;
-}
-
-namespace {
-
-QString pythonTrustSettingKeyNew(const std::string& pathUtf8)
-{
-  const QByteArray pathUtf8Bytes(pathUtf8.data(), static_cast<int>(pathUtf8.size()));
-  return QStringLiteral("python_hash/%1").arg(QString::fromUtf8(pathUtf8Bytes.toPercentEncoding()));
-}
-
-// Legacy (pre percent-encoded key): "python_hash/<local8-bit path>" (trust file id from toLocal8Bit()).
-QString pythonTrustSettingKeyLegacyLocal(const std::string& pathUtf8)
-{
-  const QString qpath =
-    QString::fromUtf8(QByteArray(pathUtf8.data(), static_cast<int>(pathUtf8.size())));
-  return QStringLiteral("python_hash/") + QString::fromLocal8Bit(qpath.toLocal8Bit());
-}
-
-// Legacy: key suffix was local 8-bit bytes from snprintf, but some builds passed char[] into
-// setValue(QString) via QString(const char*), which decodes as UTF-8 (not fromLocal8Bit).
-QString pythonTrustSettingKeyLegacyCharCtorUtf8(const std::string& pathUtf8)
-{
-  const QString qpath =
-    QString::fromUtf8(QByteArray(pathUtf8.data(), static_cast<int>(pathUtf8.size())));
-  return QStringLiteral("python_hash/") + QString::fromUtf8(qpath.toLocal8Bit());
-}
-
-// Legacy: "python_hash/<raw UTF-8 path>" (brief use of toStdString()/snprintf before UTF-8 key change).
-QString pythonTrustSettingKeyLegacyRawUtf8(const std::string& pathUtf8)
-{
-  return QStringLiteral("python_hash/") +
-         QString::fromUtf8(QByteArray(pathUtf8.data(), static_cast<int>(pathUtf8.size())));
-}
-
-QString readPythonTrustHash(QSettingsCached& settings, const std::string& pathUtf8)
-{
-  const QString kNew = pythonTrustSettingKeyNew(pathUtf8);
-  if (settings.contains(kNew)) {
-    return settings.value(kNew).toString();
-  }
-  const QString kLegLoc = pythonTrustSettingKeyLegacyLocal(pathUtf8);
-  if (settings.contains(kLegLoc)) {
-    const QString v = settings.value(kLegLoc).toString();
-    settings.setValue(kNew, v);
-    settings.remove(kLegLoc);
-    return v;
-  }
-  const QString kLegCharUtf8 = pythonTrustSettingKeyLegacyCharCtorUtf8(pathUtf8);
-  if (kLegCharUtf8 != kNew && kLegCharUtf8 != kLegLoc && settings.contains(kLegCharUtf8)) {
-    const QString v = settings.value(kLegCharUtf8).toString();
-    settings.setValue(kNew, v);
-    settings.remove(kLegCharUtf8);
-    return v;
-  }
-  const QString kLegRaw = pythonTrustSettingKeyLegacyRawUtf8(pathUtf8);
-  if (kLegRaw != kNew && kLegRaw != kLegLoc && kLegRaw != kLegCharUtf8 && settings.contains(kLegRaw)) {
-    const QString v = settings.value(kLegRaw).toString();
-    settings.setValue(kNew, v);
-    settings.remove(kLegRaw);
-    return v;
-  }
-  return {};
-}
-
-void writePythonTrustHash(QSettingsCached& settings, const std::string& pathUtf8,
-                          const std::string& hash)
-{
-  const QString kNew = pythonTrustSettingKeyNew(pathUtf8);
-  const QString v = QString::fromStdString(hash);
-  settings.setValue(kNew, v);
-  settings.remove(pythonTrustSettingKeyLegacyLocal(pathUtf8));
-  const QString kLegCharUtf8 = pythonTrustSettingKeyLegacyCharCtorUtf8(pathUtf8);
-  if (kLegCharUtf8 != kNew && kLegCharUtf8 != pythonTrustSettingKeyLegacyLocal(pathUtf8)) {
-    settings.remove(kLegCharUtf8);
-  }
-  const QString kLegRaw = pythonTrustSettingKeyLegacyRawUtf8(pathUtf8);
-  if (kLegRaw != kNew) {
-    settings.remove(kLegRaw);
-  }
-}
-
-}  // namespace
-
 static size_t curl_download_write(void *ptr, size_t size, size_t nmemb, void *stream)
 {
   QFile *fh = (QFile *)stream;
@@ -1821,9 +1713,11 @@ void MainWindow::on_fileActionPythonRevoke_triggered()
 {
   QSettingsCached settings;
 #ifdef ENABLE_PYTHON
+/*
   python_trusted = false;
   this->trusted_edit_document_name = "";
   this->untrusted_edit_document_name = "";
+*/
 #endif
   settings.remove("python_hash");
   QMessageBox::information(this, _("Trusted Files"), "All trusted python files revoked",
@@ -1832,27 +1726,7 @@ void MainWindow::on_fileActionPythonRevoke_triggered()
 
 void MainWindow::on_fileActionPythonTrustCurrent_triggered()
 {
-#ifdef ENABLE_PYTHON
-  if (activeEditor->language != LANG_PYTHON) {
-    QMessageBox::information(this, _("Python"), _("The active document is not a Python file."));
-    return;
-  }
-  if (activeEditor->filepath.isEmpty()) {
-    QMessageBox::information(
-      this, _("Python"),
-      _("Untitled buffers are already trusted. Save to a file if you need a persistent trust entry."));
-    return;
-  }
-  const QByteArray docUtf8 = activeEditor->toPlainText().toUtf8();
-  const std::string content(docUtf8.constData(), static_cast<size_t>(docUtf8.size()));
-  QSettingsCached settings;
-  const QByteArray pathUtf8 = activeEditor->filepath.toUtf8();
-  const std::string fpath(pathUtf8.constData(), static_cast<size_t>(pathUtf8.size()));
-  clearPythonUntrustStateForPath(fpath);
-  writePythonTrustHash(settings, fpath, SHA256HashString(content));
-  trusted_edit_document_name = fpath;
-  QMessageBox::information(this, _("Python"), _("This document is now trusted for Python execution."));
-#endif
+  activeEditor->trustCurrent();
 }
 
 void MainWindow::on_fileActionPythonCreateVenv_triggered()
@@ -2243,82 +2117,6 @@ bool MainWindow::fileChangedOnDisk()
    Returns true if anything was compiled.
  */
 
-#ifdef ENABLE_PYTHON
-bool MainWindow::trust_python_file(const std::string& file, const std::string& content)
-{
-  QSettingsCached settings;
-  if (python_trusted) return true;
-  if (Settings::SettingsPython::globalTrustPython.value() == true) return true;
-
-  // Trust unsaved files (empty filepath) - they're created by the user, not loaded from disk
-  if (file.empty()) {
-    return true;
-  }
-
-  std::string act_hash, ref_hash;
-  act_hash = SHA256HashString(content);
-
-  if (file == this->untrusted_edit_document_name) return false;
-
-  if (file == this->trusted_edit_document_name) {
-    writePythonTrustHash(settings, file, act_hash);
-    return true;
-  }
-
-  if (content.size() <= 1) {  // 1st character already typed
-    this->trusted_edit_document_name = file;
-    return true;
-  }
-  /*
-    // Disabled: PythonSCAD relies on a hash-based trust store (see
-    // readPythonTrustHash / writePythonTrustHash) instead of a content
-    // sniff.  Kept for historical reference; if anyone re-enables this
-    // shortcut, the prefix check must accept all three module names that
-    // a PythonSCAD script can legally start with.
-    if (content.rfind("from openscad import", 0) == 0 ||
-        content.rfind("from pythonscad import", 0) == 0 ||
-        content.rfind("from _openscad import", 0) == 0) {
-      this->trusted_edit_document_name = file;
-      return true;
-    }
-  */
-
-  ref_hash = readPythonTrustHash(settings, file).toStdString();
-
-  if (act_hash == ref_hash) {
-    this->trusted_edit_document_name = file;
-    return true;
-  }
-
-  auto ret = QMessageBox::warning(this, "Application",
-                                  _("Python files can potentially contain harmful stuff.\n"
-                                    "Do you trust this file ?\n"),
-                                  QMessageBox::Yes | QMessageBox::YesAll | QMessageBox::No);
-  if (ret == QMessageBox::YesAll) {
-    python_trusted = true;
-    return true;
-  }
-  if (ret == QMessageBox::Yes) {
-    this->trusted_edit_document_name = file;
-    writePythonTrustHash(settings, file, act_hash);
-    return true;
-  }
-
-  if (ret == QMessageBox::No) {
-    this->untrusted_edit_document_name = file;
-    return false;
-  }
-  return false;
-}
-
-void MainWindow::clearPythonUntrustStateForPath(const std::string& path)
-{
-  if (path == untrusted_edit_document_name) {
-    untrusted_edit_document_name.clear();
-  }
-}
-#endif  // ifdef ENABLE_PYTHON
-
 std::shared_ptr<SourceFile> MainWindow::parseDocument(EditorInterface *editor,
                                                       bool pythonDryRunFullScript)
 {
@@ -2338,7 +2136,7 @@ std::shared_ptr<SourceFile> MainWindow::parseDocument(EditorInterface *editor,
   auto fulltext_py = std::string(documentUtf8.constData(), static_cast<size_t>(documentUtf8.size()));
   SourceFile *sourceFile = nullptr;
 #ifdef ENABLE_PYTHON
-  if (editor->language == LANG_PYTHON && !trust_python_file(trustPathId, fulltext_py)) {
+  if (editor->language == LANG_PYTHON && !activeEditor->trust_python_file()) {
     LOG(message_group::Warning, Location::NONE, "", "Python file is not trusted");
     editor->resetHighlighting();
     editor->parameterWidget->setEnabled(false);
